Optimize

+++ src/Game/Duel/Duel.php
@@ -49,7 +49,7 @@ final class Duel extends AbstractListener
         $this->player1 = $player1;
         $this->player2 = $player2;
         foreach ([$player1, $player2] as $players) {
-            $session = PracticeCore::getSessionManager()::getSession($players);
+            $session = PracticeCore::getSessionManager()->getSession($players);
             $session->DuelClass = $this;
         }
     }
@@ -64,7 +64,7 @@ final class Duel extends AbstractListener
         $player = $event->getEntity();
         $damager = $event->getDamager();
         if ($damager instanceof Player && $player instanceof Player && $this->kit->getName() === 'Boxing' && $damager->getWorld() === $this->world) {
-            $session = PracticeCore::getSessionManager()::getSession($damager);
+            $session = PracticeCore::getSessionManager()->getSession($damager);
             $session->BoxingPoint++;
             if ($session->BoxingPoint > 99) {
                 $player->kill();
@@ -80,21 +80,19 @@ final class Duel extends AbstractListener
     {
         if ($tick % 5 === 0) {
             foreach ($this->getPlayers() as $player) {
-                $session = PracticeCore::getSessionManager()::getSession($player);
-                if ($player->isOnline()) {
-                    if ($session->isDueling) {
-                        if ($player->getWorld() === $this->world && ($player->getPosition()->getY() < 98)) {
-                            $player->kill();
-                        }
-                    } else {
-                        $this->loser = $player;
-                        $this->winner = $player->getName() !== $this->player1->getName() ? $this->player1 : $this->player2;
-                        $this->onEnd();
-                    }
-                } else {
+                $session = PracticeCore::getSessionManager()->getSession($player);
+                if (!$player->isOnline()) {
                     $this->loser = $player;
                     $this->winner = $player->getName() !== $this->player1->getName() ? $this->player1 : $this->player2;
                     $this->onEnd($player);
+                    continue;
+                }
+                if ($session->isDueling && $player->getWorld() === $this->world && $player->getPosition()->getY() < 98) {
+                    $player->kill();
+                } else {
+                    $this->loser = $player;
+                    $this->winner = $player->getName() !== $this->player1->getName() ? $this->player1 : $this->player2;
+                    $this->onEnd();
                 }
             }
         }
@@ -129,12 +127,12 @@ final class Duel extends AbstractListener
                     }
                 }
                 $this->startSec--;
-            } else {
-                if ($this->time <= 0) {
-                    $this->onEnd();
-                }
-                $this->time--;
+                return;
             }
+            if ($this->time <= 0) {
+                $this->onEnd();
+            }
+            $this->time--;
         }
     }
 
@@ -157,7 +155,7 @@ final class Duel extends AbstractListener
                 if ($playerLeft === null || $online->getName() !== $playerLeft->getName()) {
                     if ($online instanceof Player) {
                         $world = Server::getInstance()->getWorldManager()->getDefaultWorld();
-                        $session = PracticeCore::getSessionManager()::getSession($online);
+                        $session = PracticeCore::getSessionManager()->getSession($online);
                         $online->sendMessage('§f-----------------------');
                         $winnerMessage = '§aWinner: §f';
                         $winnerMessage .= $this->winner !== null ? $this->winner->getName() : 'None';
@@ -174,12 +172,12 @@ final class Duel extends AbstractListener
                         $session->DuelClass = null;
                         $session->setOpponent(null);
                         if ($this->winner !== null) {
-                            $WinnerSession = PracticeCore::getSessionManager()::getSession($this->winner);
+                            $WinnerSession = PracticeCore::getSessionManager()->getSession($this->winner);
                             $WinnerSession->kills++;
                             $WinnerSession->killStreak++;
                         }
                         if ($this->loser !== null) {
-                            $LoserSession = PracticeCore::getSessionManager()::getSession($this->loser);
+                            $LoserSession = PracticeCore::getSessionManager()->getSession($this->loser);
                             $LoserSession->deaths++;
                             $LoserSession->killStreak = 0;
                         }
